--- Source/FileHashChecker.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30406.217
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FileHashChecker", "FileHashChecker\FileHashChecker.csproj", "{E26B3800-AEA3-4564-8868-F49F82A1C105}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IconImage", "IconImage\IconImage.csproj", "{FF649D56-8E56-4A34-812B-0455D1A4A616}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E26B3800-AEA3-4564-8868-F49F82A1C105}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E26B3800-AEA3-4564-8868-F49F82A1C105}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E26B3800-AEA3-4564-8868-F49F82A1C105}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E26B3800-AEA3-4564-8868-F49F82A1C105}.Release|Any CPU.Build.0 = Release|Any CPU
-		{FF649D56-8E56-4A34-812B-0455D1A4A616}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FF649D56-8E56-4A34-812B-0455D1A4A616}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FF649D56-8E56-4A34-812B-0455D1A4A616}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FF649D56-8E56-4A34-812B-0455D1A4A616}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F47489B7-FB68-4914-A7C3-7C9B42B8A6E0}
-	EndGlobalSection
-EndGlobal
