--- Source/FileHashChecker/App.xaml.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Configuration;
-using System.Data;
-using System.Linq;
-using System.Threading.Tasks;
-using System.Windows;
-
-namespace FileHashChecker
-{
-	public partial class App : Application
-	{
-	}
-}
