--- Source/FileHashChecker/Views/Converters/BooleanInverseConverter.cs
@@ -1,29 +1,0 @@
-﻿using System;
-using System.Globalization;
-using System.Windows;
-using System.Windows.Data;
-
-namespace FileHashChecker.Views.Converters
-{
-	[ValueConversion(typeof(bool), typeof(bool))]
-	public class BooleanInverseConverter : IValueConverter
-	{
-		public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-		{
-			return ConvertBase(value);
-		}
-
-		public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-		{
-			return ConvertBase(value);
-		}
-
-		private static object ConvertBase(object value)
-		{
-			if (!(value is bool sourceValue))
-				return DependencyProperty.UnsetValue;
-
-			return !sourceValue;
-		}
-	}
-}
