+++ Source/HashPad/App.config
