+++ Source/HashPad/App.xaml
@@ -1,0 +1,8 @@
+﻿<Application x:Class="HashPad.App"
+			 xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+			 xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+			 StartupUri="Views/MainWindow.xaml">
+	<Application.Resources>
+
+	</Application.Resources>
+</Application>
