+++ Source/HashPad/App.xaml.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Configuration;
+using System.Data;
+using System.Linq;
+using System.Threading.Tasks;
+using System.Windows;
+
+namespace HashPad
+{
+	public partial class App : Application
+	{
+	}
+}
