+++ Source/HashPad/Common/NotificationObject.cs
@@ -1,0 +1,23 @@
+﻿using System.Collections.Generic;
+using System.ComponentModel;
+using System.Runtime.CompilerServices;
+
+namespace HashPad.Common
+{
+	public abstract class NotificationObject : INotifyPropertyChanged
+	{
+		public event PropertyChangedEventHandler PropertyChanged;
+
+		protected void SetProperty<T>(ref T storage, T value, [CallerMemberName] string propertyName = null)
+		{
+			if (EqualityComparer<T>.Default.Equals(storage, value))
+				return;
+
+			storage = value;
+			RaisePropertyChanged(propertyName);
+		}
+
+		protected void RaisePropertyChanged([CallerMemberName] string propertyName = null) =>
+			PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+	}
+}
