+++ Source/HashPad/Properties/Resources.resx
