+++ Source/HashPad/Properties/Settings.settings
