+++ Source/HashPad/Resources/Hash.ico
